+++ ChangeLog.md
@@ -13,6 +13,7 @@
     - Use pwix:logger universal logger
     - Fix array concatenation of Entities, Records and Tabular fieldsets
     - Fix stylesheet to be compatible with pwix:validity flex display
+    - Protect UI against consequences of removed validity period
 
 ### 1.5.0